Added my fortnite b checker, is not hooked up to the bot
